refactor tests to a package
