+++ .gitignore
@@ -1,3 +1,7 @@
-*.pyc
-*~
 build
+
+# cache files
+*~
+*.pyc
+*.swo
+*.swp
